refactor: split app bootstrap and window event hooks
